`$top` command reintegrated as `$stats top-chatters`
